app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-logging==0.50b0
prometheus-client==0.21.1
httpx==0.28.1
